backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]
aiosqlite
sqlalchemy[asyncio]
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
google-genai>=1.0.0
python-dotenv
pydantic>=2.0
pydantic-settings
httpx
mcp
